AtlasSheet now uses float for spriteWidth and spriteHeight. Updated Platformer2D demo to use new atlasing system.

+++ Drivers/D3D11Renderer/Include/D3D11SpriteMaterial.h
@@ -34,8 +34,8 @@ namespace Monsoon {
 
 			struct AtlasBuffer
 			{
-				int srcWidth;
-				int srcHeight;
+				float srcWidth;
+				float srcHeight;
 				float srcU;
 				float srcV;
 			};
